GUACAMOLE-5: Move permission management to own resource.
